give more commands their own editor
