+++ src/frontend/components/Commands/DictLookupCommandEditor.vue
@@ -32,34 +32,40 @@
                 </div>
               </td>
             </tr>
-            <tr>
-              <td>Response:</td>
-              <td>
-                <div v-for="(txt, idx) in item.data.text" :key="idx" class="field textarea-holder">
-                  <textarea class="textarea" type="text" v-model="item.data.text[idx]" :class="{
-                    'has-background-danger-light': !item.data.text[idx],
-                    'has-text-danger-dark': !item.data.text[idx],
-                  }" />
+            <template>
+              <tr>
+                <td>Language:</td>
+                <td>
+                  <input class="input is-small spaceinput mb-1" v-model="item.data.lang" />
+                  <span v-for="(lang, idx) in dictLangs" :key="idx" class="button is-small mr-1"
+                    @click="item.data.lang = lang.value" :title="lang.title">{{ lang.flag }}</span>
+                  <span class="button is-small mr-1" @click="item.data.lang = '$args(0)'"><code>$args(0)</code></span>
+                </td>
+              </tr>
+              <tr>
+                <td>Phrase:</td>
+                <td>
+                  <input class="input is-small spaceinput mb-1" v-model="item.data.phrase" />
+                  <span class="button is-small mr-1" @click="item.data.phrase = ''">All args</span>
+                  <span class="button is-small mr-1"
+                    @click="item.data.phrase = '$args(1:)'"><code>$args(1:)</code></span>
+                </td>
+              </tr>
+              <tr>
+                <td>Response:</td>
+                <td>
                   <div class="help">
-                    <macro-select @selected="insertMacro(idx, $event)" />
+                    Outputs the translation for the input phrase. The
+                    translation is always from/to english. <br />
+                    To let the user decide on the language use
+                    <code>$args(0)</code> as language, and
+                    <code>$args(1:)</code> as phrase. <br />
+                    If phrase is left empty, all arguments to the command will
+                    be used as the phrase.
                   </div>
-                  <button class="button is-small" :disabled="item.data.text.length <= 1" @click="rmtxt(idx)">
-                    <i class="fa fa-remove" />
-                  </button>
-                </div>
-                <div class="field">
-                  <button class="button is-small" @click="addtxt">
-                    <i class="fa fa-plus mr-1" /> Add response
-                  </button>
-                </div>
-                <div>
-                  <p class="help">
-                    If multiple responses exist, a random one will be used when
-                    the command is triggered.
-                  </p>
-                </div>
-              </td>
-            </tr>
+                </td>
+              </tr>
+            </template>
             <tr>
               <td>Variables:</td>
               <td>
@@ -170,7 +176,6 @@ import { permissions } from "../../../common/permissions";
 import {
   commands,
   isValidTrigger,
-  newText,
   newTrigger,
 } from "../../../common/commands";
 import {
@@ -178,13 +183,18 @@ import {
   CommandTriggerType,
   CommandVariable,
   CommandVariableChange,
+  DictLookupCommand,
   GlobalVariable,
-  RandomTextCommand,
 } from "../../../types";
 
 interface AutocompletableVariable {
   var: CommandVariable | GlobalVariable;
   type: string;
+}
+interface ComponentDataLang {
+  value: string;
+  flag: string;
+  title: string;
 }
 
 interface ComponentDataPermission {
@@ -193,15 +203,16 @@ interface ComponentDataPermission {
 }
 
 interface ComponentData {
-  item: RandomTextCommand | null;
+  item: DictLookupCommand | null;
   variableChangeFocusIdx: number;
+  dictLangs: ComponentDataLang[];
   possiblePermissions: ComponentDataPermission[];
 }
 
 export default defineComponent({
   props: {
     modelValue: {
-      type: Object as PropType<RandomTextCommand>,
+      type: Object as PropType<DictLookupCommand>,
       required: true,
     },
     mode: {
@@ -221,6 +232,15 @@ export default defineComponent({
   data: (): ComponentData => ({
     item: null,
     variableChangeFocusIdx: -1,
+    dictLangs: [
+      { value: "ja", flag: "🇯🇵", title: "Japanese" },
+      { value: "ru", flag: "🇷🇺", title: "Russian" },
+      { value: "de", flag: "🇩🇪", title: "German" },
+      { value: "es", flag: "🇪🇸", title: "Spanish" },
+      { value: "fr", flag: "🇫🇷", title: "French" },
+      { value: "it", flag: "🇮🇹", title: "Italian" },
+      { value: "pt", flag: "🇵🇹/🇧🇷", title: "Portuguese" },
+    ],
     possiblePermissions: permissions,
   }),
   mounted() {
@@ -238,13 +258,6 @@ export default defineComponent({
     },
   },
   methods: {
-    addtxt(): void {
-      if (!this.item) {
-        console.warn("addtxt: this.item not initialized");
-        return;
-      }
-      this.item.data.text.push(newText());
-    },
     addtrigger(trigger: any): void {
       if (!this.item) {
         console.warn("addtrigger: this.item not initialized");
@@ -303,15 +316,6 @@ export default defineComponent({
     onOverlayClick(): void {
       this.$emit("cancel");
     },
-    rmtxt(idx: number): void {
-      if (!this.item) {
-        console.warn("rmtxt: this.item not initialized");
-        return;
-      }
-      this.item.data.text = this.item.data.text.filter(
-        (_val: string, index: number) => index !== idx
-      );
-    },
     rmtrigger(idx: number): void {
       if (!this.item) {
         console.warn("rmtrigger: this.item not initialized");
@@ -320,13 +324,6 @@ export default defineComponent({
       this.item.triggers = this.item.triggers.filter(
         (_val: CommandTrigger, index: number) => index !== idx
       );
-    },
-    insertMacro(idx: number, macro: { value: string; title: string }): void {
-      if (!this.item) {
-        console.warn("insertMacro: this.item not initialized");
-        return;
-      }
-      this.item.data.text[idx] += macro.value;
     },
     autocompletableVariables(): AutocompletableVariable[] {
       if (!this.item) {
@@ -383,13 +380,6 @@ export default defineComponent({
         }
       }
 
-      // check if settings are correct
-      for (const t of this.item.data.text) {
-        if (t === "") {
-          return false;
-        }
-      }
-
       return true;
     },
     actionDescription(): string {
@@ -411,15 +401,3 @@ export default defineComponent({
   },
 });
 </script>
-<style scoped>
-.textarea-holder {
-  position: relative;
-  padding-right: 2em;
-}
-
-.textarea-holder .button {
-  position: absolute;
-  right: -2px;
-  top: 0;
-}
-</style>
